% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA_from_file.R
\name{CCA_from_file}
\alias{CCA_from_file}
\title{Do Canonical correlation analysis}
\usage{
CCA_from_file(file1, file2, rank, pval_method)
}
\arguments{
\item{file1}{file for A matrix (.txt file)}

\item{file2}{file for B matrix}

\item{rank}{rank for the decompostiton}

\item{pval_method}{default = "Wilks". Method for testing signficance of components (Wilks, Hotelling,Pillai,Roy)}

\item{standardize}{default = T}
}
\description{
Input: matrix A (ex. rows as genes, cols as samples)
Input: matrix B (ex. rows as genes, cols as samples)
Output: writes to file scores and loadings of decomposed matrices that maximize the correlation between A and B, as well as p-values for each component
}
\details{
WARNING - script will internally convert all colnames to alphanumeric only.

Returns cca object, number of significant canonical dimensions (can be smaller than #of vars in smaller set)
}
