% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_kmeans_clustering.R
\name{do_kmeans_clustering}
\alias{do_kmeans_clustering}
\title{Do k-means clustering on data matrix, genes in rownames, samples in columnnames
Input: Numeric only data matrix with gene names in rownames
Output: prints clustered heatmap}
\usage{
do_kmeans_clustering(data_matrix, k_clusters = 5, colseps = NULL,
  annotation = NULL, breaks = c(-4, seq(-1.5, 1.5, length = 100), 4))
}
\arguments{
\item{data_matrix}{matrix or dataframe to cluster, numeric only}

\item{k_clusters}{cluster number, default = 5}

\item{colseps}{vector of numbers for column separation gaps}

\item{annotation}{Annotation data frame with genenames in rownames}
}
\description{
Do k-means clustering on data matrix, genes in rownames, samples in columnnames
Input: Numeric only data matrix with gene names in rownames
Output: prints clustered heatmap
}
