% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_squared.R
\name{gsea_squared}
\alias{gsea_squared}
\title{#GSEA-squared (last mod. 07/2019 ksheu)}
\usage{
gsea_squared(file1, file2, keywords, keyword.labels = keywords,
  get.keywords = F, dir = getwd())
}
\arguments{
\item{file1, file2}{Paths to GSEA excel file outputs}

\item{keywords}{Keyword groups to search for (eg. c("CELL_CYLE", "INFLAMMATORY|IMMUNE"))}

\item{keyword.labels}{Labels for keyword groups, in the same order (eg. c("cycle", "immune"))}

\item{get.keywords}{perform ks tests on individual words to get candidate keywords}

\item{dir}{set output directory}
}
\description{
Input: excel file paths for GSEA output, pos and neg enrichment
Input: keyword groups to search 
Output: graph and pvals
}
