% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA_from_file.R
\name{ICA_from_file}
\alias{ICA_from_file}
\title{Do Independent components analysis
The input matrix X (input as genes in rows, samples in columns) is a linear combination of non-Gaussian independent components.
As opposed to PCA, which assumes generally Gaussian distributed data.
Assume X = SA, where S are the independent components.
We want to find another matrix W , where XW = S, to unmix}
\usage{
ICA_from_file(data, data.labels, comps = 5)
}
\arguments{
\item{data}{file for the input matrix X (.txt or .RDS file)}

\item{data.labels}{labels for samples with two columns: samp.names and type}

\item{comps}{number of components to be extracted, defualt = 5}
}
\description{
Input: matrix X (ex. rows as genes, cols as samples)
Output: plots only currently
}
