% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_PCA_from_file.R
\name{log_PCA_from_file}
\alias{log_PCA_from_file}
\title{PCA from file after taking log2}
\usage{
log_PCA_from_file(file, center = TRUE, scale = FALSE)
}
\arguments{
\item{file}{Filepath/filename of data matrix with no row numbering}

\item{center}{default=T}

\item{scale}{default=F}
}
\description{
Reads file with samples in columns and variables in rows. Takes log2(x+1) and does PCA. Writes to file scores, loadings, eigenvalues.
}
