% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_loadings_gctfile.R
\name{make_loadings_gctfile}
\alias{make_loadings_gctfile}
\title{Format PCA loadings for IGV}
\usage{
make_loadings_gctfile(file, ncomps = 5)
}
\arguments{
\item{file}{loadings file (output of PCA_from_file)}

\item{ncomps}{number of components}
}
\description{
Outputs gct file in format for IGV
}

