% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rnk_file.R
\name{make_rnk_file}
\alias{make_rnk_file}
\title{Make .rnk file
Input: txt file
Output: writes out .rnk file}
\usage{
make_rnk_file(file, column)
}
\arguments{
\item{file}{filename of some txt file}

\item{column}{column number of orig file you want to rank (>1)}
}
\description{
make .rnk file (ie. for GSEA preranked)
}
