% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scores_file.R
\name{make_scores_file}
\alias{make_scores_file}
\title{Makes scores file for IGV}
\usage{
make_scores_file(file, ncomps = 5)
}
\arguments{
\item{file}{Score file (output of PCA_from_file)}

\item{ncomps}{number of PCs wanted}
}
\description{
Writes out minimized scores file
}

