% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNMF_from_file.R
\name{NNMF_from_file}
\alias{NNMF_from_file}
\title{Nonnegative matrix factorization}
\usage{
NNMF_from_file(file, rank = 5, method = "scd", loss = "mkl", impute = F,
  verbose = 2)
}
\arguments{
\item{file}{file for A matrix (.txt or .RDS file)}

\item{rank}{rank for the decompostiton}

\item{method}{'scd' (seq. coord-wise descent) or 'lee' (Lee's algo)}

\item{loss}{'mse' (mean squared error) or 'mkl' (mean KL divergence) 
for sc data probably better to use KL divergence? (Frobenius norm is standard)}

\item{impute}{default=F Simple NNMF or impute 0 values by first setting them all to NA}

\item{verbose}{0-2}
}
\description{
Input: matrix A (ex. rows as genes, cols as samples)
Output: writes to file decomposed matrices W (loadings) and H (scores) that minimize the loss between the original and decomposed matrices
}
\details{
Returns nnmf object
}

