% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_and_save.R
\name{plot_pca_and_save}
\alias{plot_pca_and_save}
\title{Plot PCA with save option}
\usage{
plot_pca_and_save(scores.file, info.name, info.type, PCx = "PC1",
  PCy = "PC2", factor.levels = levels(info.type), aes_fill = NULL,
  savename, savePlot = F, width = 6, height = 2, title = "",
  label = F, savetype = ".pdf", w = 6, h = 3, aspect.ratio = 0.8,
  legend = F, legendname = "default", colors = colpalette[1:4],
  psize = 2, plotLabelname = "none")
}
\arguments{
\item{scores.file}{PCA score file (scores file, output of PCA_from_file)}

\item{info.name}{vector of sample names}

\item{info.type}{vector of sample types in same order as names}

\item{PCx, PCy}{PCs to display}

\item{factor.levels}{ie. levels <- c("Undifferentiated", "Neural crest-like", "Transitory", "Melanocytic"))}

\item{aes_fill}{column header to color samples by}

\item{savename}{prefix to save file by}

\item{savePlot}{default=F}

\item{width, height}{size of the plot}

\item{title}{title of the plot}

\item{label}{column name to label samples by on PCA plot, default = F}

\item{savetype}{".pdf" or ".png" or ".tiff", etc.}

\item{colors}{ie. colpalette <- c("Cluster 1"="#4F81BD","Cluster 2"="#C0504D","Cluster 3"="#8064A2","Cluster 4"="#9BBB59")}
}
\description{
Plot PCA with save option
}
