% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_projection_only.R
\name{plot_pca_projection_only}
\alias{plot_pca_projection_only}
\title{Plot PCA projection only}
\usage{
plot_pca_projection_only(rotated.file2, info.name, info.type,
  title = "Projection", labels = TRUE, PCx = "PC1", PCy = "PC2")
}
\arguments{
\item{rotated.file2}{File containing scores matrix}

\item{info.name}{Vector of sample names}

\item{info.type}{Vector of sample types in the same order}

\item{title}{Title of the plot}

\item{labels}{default=T}

\item{PCx, PCy}{PCs to display}
}
\description{
Plots projection from rotated.scores file (output of intersect_do_PCA_and_project_second_dataset)
}

