% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA_from_file.R
\name{PLSDA_from_file}
\alias{PLSDA_from_file}
\title{Partial Least Squares from file}
\usage{
PLSDA_from_file(file, sample.names, response.values, comps = 2,
  scale = F, ind.names = F)
}
\arguments{
\item{file}{file for X matirx}

\item{sample.names}{Vector of sample names in X matrix}

\item{response.values}{Vector of response values in same order matching sample.names}

\item{scale}{default=T}

\item{comp}{number of components to compute}
}
\description{
Writes out loadings
}
