% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomp.R
\name{tensor_decomp}
\alias{tensor_decomp}
\title{Implements tensor factorization methods, Tucker and PARAFAC/canonical polyadic decomp}
\usage{
tensor_decomp(arr, method, rank = tnsr@modes)
}
\arguments{
\item{arr}{multidimensional array that get converted through as.tensor(arr)}

\item{method}{"tucker" or "cp"}

\item{rank}{default = 3, number (if CP) or vector of numbers (if tucker) that is <= the ranks of the tensor}
}
\description{
Tucker:  If there is no truncationin one of the modes, then this is the same as the MPCA,mpca.  
If there is no truncation in all the modes, then this is the same as the HOSVD,hosvd.
}
\details{
Input: Multidimensional array that gets converted to as.tensor
Output: Decomposed tensor object
}
